[*] реорганизация
